Move some server pieces from main-command to main
