+++ main-command/src/main/scala/sbt/Command.scala
@@ -88,21 +88,6 @@ object Command {
       }
     }
 
-  /** This is the main function State transfer function of the sbt command processing, called by MainLoop.next, */
-  def process(exec: Exec, state: State): State =
-    {
-      val channelName = exec.source map { _.channelName }
-      State.exchange.publishEventMessage(ExecStatusEvent("Processing", channelName, exec.execId, Vector()))
-      val parser = combine(state.definedCommands)
-      val newState = parse(exec.commandLine, parser(state)) match {
-        case Right(s) => s() // apply command.  command side effects happen here
-        case Left(errMsg) =>
-          state.log.error(errMsg)
-          state.fail
-      }
-      State.exchange.publishEventMessage(ExecStatusEvent("Done", channelName, exec.execId, newState.remainingCommands.toVector map { _.commandLine }))
-      newState
-    }
   def invalidValue(label: String, allowed: Iterable[String])(value: String): String =
     "Not a valid " + label + ": " + value + similar(value, allowed)
   def similar(value: String, allowed: Iterable[String]): String =
