+++ main/src/main/scala/sbt/internal/server/NetworkChannel.scala
